refactor: rename controllers/http to controllers/rest for better API semantics
  - Rename http/ directory to rest/ to better reflect RESTful API implementation
  - Update import statements in main.py to use new rest module paths
  - Rename test files to align with new naming convention
  - Update documentation and comments to use consistent REST terminology
