+++ src/controllers/rest/pokemon/mapper.py
