+++ src/controllers/rest/pokemon/schema.py
